feat: add HomePageController (#20)
