+++ lib/app/bindings/main_binding.dart
@@ -1,4 +1,5 @@
 import 'package:get/get.dart';
+import 'package:one_day_one_something/app/controller/main/home_page_controller.dart';
 import 'package:one_day_one_something/app/controller/main/main_controller.dart';
 import 'package:one_day_one_something/app/controller/main/setting_controller.dart';
 
@@ -7,6 +8,16 @@ class MainBinding implements Bindings {
   void dependencies() {
     Get.lazyPut<MainController>(() {
       return MainController();
+    });
+    HomePageBinding().dependencies();
+  }
+}
+
+class HomePageBinding implements Bindings{
+  @override
+  void dependencies() {
+    Get.lazyPut<HomePageController>(() {
+      return HomePageController();
     });
   }
 }
